Formatting VaultPriceOracle

+++ src/core/VaultPriceOracle.sol
@@ -7,7 +7,6 @@ import {Controllable, IControllable} from "./base/Controllable.sol";
 /// @author ruby (https://github.com/alexandersazonof)
 /// @dev A contract for aggregating vault prices from multiple validators using a quorum-based median mechanism.
 contract VaultPriceOracle is Controllable, IVaultPriceOracle {
-
     /*´:°•.°+.*•´.*:˚.°*.˚•´.°:°•.°•.*•´.*:˚.°*.˚•´.°:°•.°+.*•´.*:*/
     /*                         CONSTANTS                          */
     /*.•°:°.´+˚.*°.˚:*.´•*.+°.•°:´*.´•*.•°.•°:°.´:•˚°.*°.˚:*.´+°.•*/
@@ -17,7 +16,7 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
 
     // keccak256(abi.encode(uint256(keccak256("erc7201:stability.VaultPriceOracle")) - 1)) & ~bytes32(uint256(0xff));
     bytes32 private constant VAULT_PRICE_ORACLE_STORAGE_LOCATION =
-    0xa68171b251d015e5a139782486873a18b874637da10a73c080418fb52ac37300;
+        0xa68171b251d015e5a139782486873a18b874637da10a73c080418fb52ac37300;
 
     /*´:°•.°+.*•´.*:˚.°*.˚•´.°:°•.°•.*•´.*:˚.°*.˚•´.°:°•.°+.*•´.*:*/
     /*                          STORAGE                           */
@@ -26,11 +25,11 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /// @custom:storage-location erc7201:stability.VaultPriceOracle
     struct VaultPriceOracleStorage {
         mapping(address => AggregatedData) vaultPrices;
-        mapping(address => mapping(uint256 => mapping(address => Observation))) observations;
+        mapping(address => mapping(uint => mapping(address => Observation))) observations;
         mapping(address validator => bool authorized) authorizedValidators;
         address[] validatorList;
-        uint256 minQuorum;
-        uint256 maxPriceAge;
+        uint minQuorum;
+        uint maxPriceAge;
     }
 
     /*´:°•.°+.*•´.*:˚.°*.˚•´.°:°•.°•.*•´.*:˚.°*.˚•´.°:°•.°+.*•´.*:*/
@@ -47,13 +46,18 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /*.•°:°.´+˚.*°.˚:*.´•*.+°.•°:´*.´•*.•°.•°:°.´:•˚°.*°.˚:*.´+°.•*/
 
     /// @inheritdoc IVaultPriceOracle
-    function initialize(address platform_, uint256 minQuorum_, address[] memory validators_, uint256 maxPriceAge_) public initializer {
+    function initialize(
+        address platform_,
+        uint minQuorum_,
+        address[] memory validators_,
+        uint maxPriceAge_
+    ) public initializer {
         VaultPriceOracleStorage storage $ = _getStorage();
         __Controllable_init(platform_);
 
         $.minQuorum = minQuorum_;
         $.maxPriceAge = maxPriceAge_;
-        for (uint256 i = 0; i < validators_.length; i++) {
+        for (uint i = 0; i < validators_.length; i++) {
             $.authorizedValidators[validators_[i]] = true;
             $.validatorList.push(validators_[i]);
         }
@@ -64,9 +68,9 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /*.•°:°.´+˚.*°.˚:*.´•*.+°.•°:´*.´•*.•°.•°:°.´:•˚°.*°.˚:*.´+°.•*/
 
     /// @inheritdoc IVaultPriceOracle
-    function submitPrice(address vault_, uint256 price_, uint256 roundId_) external onlyValidator {
+    function submitPrice(address vault_, uint price_, uint roundId_) external onlyValidator {
         VaultPriceOracleStorage storage $ = _getStorage();
-        uint256 currentRoundId = $.vaultPrices[vault_].roundId == 0 ? 1 : $.vaultPrices[vault_].roundId;
+        uint currentRoundId = $.vaultPrices[vault_].roundId == 0 ? 1 : $.vaultPrices[vault_].roundId;
         require(roundId_ == currentRoundId, "Invalid roundId");
 
         $.observations[vault_][currentRoundId][msg.sender] = Observation(price_, block.timestamp);
@@ -91,7 +95,7 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
         VaultPriceOracleStorage storage $ = _getStorage();
         require($.authorizedValidators[validator_], "Validator not authorized");
         $.authorizedValidators[validator_] = false;
-        for (uint256 i = 0; i < $.validatorList.length; i++) {
+        for (uint i = 0; i < $.validatorList.length; i++) {
             if ($.validatorList[i] == validator_) {
                 $.validatorList[i] = $.validatorList[$.validatorList.length - 1];
                 $.validatorList.pop();
@@ -105,9 +109,8 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /*                      VIEW FUNCTIONS                        */
     /*.•°:°.´+˚.*°.˚:*.´•*.+°.•°:´*.´•*.•°.•°:°.´:•˚°.*°.˚:*.´+°.•*/
 
-
     /// @inheritdoc IVaultPriceOracle
-    function getLatestPrice(address vault_) external view returns (uint256 price, uint256 timestamp, uint256 roundId) {
+    function getLatestPrice(address vault_) external view returns (uint price, uint timestamp, uint roundId) {
         VaultPriceOracleStorage storage $ = _getStorage();
         AggregatedData memory data = $.vaultPrices[vault_];
         require(data.timestamp > 0, "No data available");
@@ -116,14 +119,18 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     }
 
     /// @inheritdoc IVaultPriceOracle
-    function vaultPrices(address vault_) external view returns (uint256 price, uint256 timestamp, uint256 roundId) {
+    function vaultPrices(address vault_) external view returns (uint price, uint timestamp, uint roundId) {
         VaultPriceOracleStorage storage $ = _getStorage();
         AggregatedData memory data = $.vaultPrices[vault_];
         return (data.price, data.timestamp, data.roundId);
     }
 
     /// @inheritdoc IVaultPriceOracle
-    function observations(address vault_, uint256 roundId_, address validator_) external view returns (uint256 price, uint256 timestamp) {
+    function observations(
+        address vault_,
+        uint roundId_,
+        address validator_
+    ) external view returns (uint price, uint timestamp) {
         VaultPriceOracleStorage storage $ = _getStorage();
         Observation memory data = $.observations[vault_][roundId_][validator_];
         return (data.price, data.timestamp);
@@ -136,26 +143,26 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     }
 
     /// @inheritdoc IVaultPriceOracle
-    function validatorList(uint256 index_) external view returns (address) {
+    function validatorList(uint index_) external view returns (address) {
         VaultPriceOracleStorage storage $ = _getStorage();
         require(index_ < $.validatorList.length, "Index out of bounds");
         return $.validatorList[index_];
     }
 
     /// @inheritdoc IVaultPriceOracle
-    function validatorListLength() external view returns (uint256) {
+    function validatorListLength() external view returns (uint) {
         VaultPriceOracleStorage storage $ = _getStorage();
         return $.validatorList.length;
     }
 
     /// @inheritdoc IVaultPriceOracle
-    function minQuorum() external view returns (uint256) {
+    function minQuorum() external view returns (uint) {
         VaultPriceOracleStorage storage $ = _getStorage();
         return $.minQuorum;
     }
 
     /// @inheritdoc IVaultPriceOracle
-    function maxPriceAge() external view returns (uint256) {
+    function maxPriceAge() external view returns (uint) {
         VaultPriceOracleStorage storage $ = _getStorage();
         return $.maxPriceAge;
     }
@@ -169,10 +176,10 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /// @param vault_ The address of the vault to check submissions for.
     /// @param roundId_ The ID of the round to count submissions for.
     /// @return count The number of valid submissions in the specified round.
-    function _countSubmissions(address vault_, uint256 roundId_) internal view returns (uint256) {
+    function _countSubmissions(address vault_, uint roundId_) internal view returns (uint) {
         VaultPriceOracleStorage storage $ = _getStorage();
-        uint256 count = 0;
-        for (uint256 i = 0; i < $.validatorList.length; i++) {
+        uint count = 0;
+        for (uint i = 0; i < $.validatorList.length; i++) {
             if ($.observations[vault_][roundId_][$.validatorList[i]].timestamp > 0) {
                 count++;
             }
@@ -185,13 +192,13 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     ///      Emits a PriceUpdated event upon successful aggregation.
     /// @param vault_ The address of the vault to aggregate prices for.
     /// @param roundId_ The ID of the round to aggregate.
-    function _aggregateAndUpdate(address vault_, uint256 roundId_) internal {
+    function _aggregateAndUpdate(address vault_, uint roundId_) internal {
         VaultPriceOracleStorage storage $ = _getStorage();
-        uint256[] memory prices = new uint256[]($.validatorList.length);
-        uint256 validCount = 0;
-        uint256 newRoundId = roundId_ + 1;
+        uint[] memory prices = new uint[]($.validatorList.length);
+        uint validCount = 0;
+        uint newRoundId = roundId_ + 1;
 
-        for (uint256 i = 0; i < $.validatorList.length; i++) {
+        for (uint i = 0; i < $.validatorList.length; i++) {
             Observation memory obs = $.observations[vault_][roundId_][$.validatorList[i]];
             if (obs.timestamp > 0) {
                 prices[validCount] = obs.price;
@@ -199,7 +206,7 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
             }
         }
 
-        uint256 medianPrice = _getMedian(prices, validCount);
+        uint medianPrice = _getMedian(prices, validCount);
 
         $.vaultPrices[vault_] = AggregatedData(medianPrice, block.timestamp, newRoundId);
         emit PriceUpdated(vault_, medianPrice, roundId_, block.timestamp);
@@ -211,10 +218,10 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /// @param prices_ The array containing all submitted prices (including unused slots).
     /// @param count_ The number of valid prices in the array.
     /// @return The median price, or 0 if no valid prices are provided.
-    function _getMedian(uint256[] memory prices_, uint256 count_) internal pure returns (uint256) {
+    function _getMedian(uint[] memory prices_, uint count_) internal pure returns (uint) {
         if (count_ == 0) return 0;
-        uint256[] memory prices = new uint256[](count_);
-        for (uint256 i = 0; i < count_; i++) {
+        uint[] memory prices = new uint[](count_);
+        for (uint i = 0; i < count_; i++) {
             prices[i] = prices_[i];
         }
         _quickSort(prices, 0, count_ - 1);
@@ -230,9 +237,9 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /// @param arr_ The array to sort.
     /// @param low_ The starting index of the subarray to sort.
     /// @param high_ The ending index of the subarray to sort.
-    function _quickSort(uint256[] memory arr_, uint256 low_, uint256 high_) internal pure {
+    function _quickSort(uint[] memory arr_, uint low_, uint high_) internal pure {
         if (low_ < high_) {
-            uint256 pi = _partition(arr_, low_, high_);
+            uint pi = _partition(arr_, low_, high_);
             if (pi > 0) {
                 _quickSort(arr_, low_, pi - 1);
             }
@@ -246,11 +253,11 @@ contract VaultPriceOracle is Controllable, IVaultPriceOracle {
     /// @param low_ The starting index of the subarray.
     /// @param high_ The ending index of the subarray (pivot location).
     /// @return The index of the pivot after partitioning.
-    function _partition(uint256[] memory arr_, uint256 low_, uint256 high_) internal pure returns (uint256) {
-        uint256 pivot = arr_[high_];
-        uint256 i = low_;
+    function _partition(uint[] memory arr_, uint low_, uint high_) internal pure returns (uint) {
+        uint pivot = arr_[high_];
+        uint i = low_;
 
-        for (uint256 j = low_; j < high_; j++) {
+        for (uint j = low_; j < high_; j++) {
             if (arr_[j] <= pivot) {
                 (arr_[i], arr_[j]) = (arr_[j], arr_[i]);
                 i++;
